se modifica clases personaje principal, objeto dinamico, movimiento circular y resorte para hacer mejoras en el front

+++ objetodinamico.cpp
@@ -5,7 +5,7 @@ double objetoDinamico::getPosx()
     return posx;
 }
 
-void objetoDinamico::setPosx(double newPosx)
+void objetoDinamico::setPosx(float newPosx)
 {
     posx = newPosx;
 }
@@ -15,7 +15,7 @@ double objetoDinamico::getPosy()
     return posy;
 }
 
-void objetoDinamico::setPosy(double newPosy)
+void objetoDinamico::setPosy(float newPosy)
 {
     posy = newPosy;
 }
@@ -37,7 +37,7 @@ objetoDinamico::objetoDinamico()
 }
 
 
-objetoDinamico::objetoDinamico(double posx,double posy,int ancho,int alto)
+objetoDinamico::objetoDinamico(float posx,float posy,int ancho,int alto)
 
 {
     this->posx=posx;
@@ -51,27 +51,27 @@ objetoDinamico::objetoDinamico(double posx,double posy,int ancho,int alto)
     setPos(posx,posy);
 
 }
-void objetoDinamico::MoveRight(double velocidad)
+void objetoDinamico::MoveRight(float velocidad)
 {
     this->posx+=velocidad;
     setPos(posx,posy);
 }
 
-void objetoDinamico::MoveLeft(double velocidad)
+void objetoDinamico::MoveLeft(float velocidad)
 {
     this->posx-=velocidad;
 
     setPos(posx,posy);
 }
 
-void objetoDinamico::MoveUp(double velocidad)
+void objetoDinamico::MoveUp(float velocidad)
 {
     this->posy-=velocidad;
 
     setPos(posx,posy);
 }
 
-void objetoDinamico::MoveDown(double velocidad)
+void objetoDinamico::MoveDown(float velocidad)
 {
     this->posy+=velocidad;
